Add synchronizer for sensor data
We need a custom synchronizer for sensor data.
This synchronizer lets us block a thread until the sensors indicate something. For example, after sending the arm command, we can block until the sensors indicate we're armed.

We need a synchronizer because the sensor data comes from one or more threads owned by rospy. But the main thread (or some other application thread) needs to respond.