Hover tests works in gazebo
Note that I had to use the same GCS heartbeat hack.

The program sends heartbeats as if it's a separate system that reports as a ground control station . Maybe there is a way to fly without  while not mimicking a GCS. At this time, the flight controller  triggers the data link loss fail safe when it doesn't see a GCS heartbeat.