WIP fly box shape test (UNTESTED CODE)
We need a test that exercises setpoint control.
That's what the box test is supposed to do. It traces out a square in the sky.

Refactor to help with the box test. The hover test and the box test share similar functionality. The functions that test if the drone has finished some task have been moved to flight_predicate.py

Change how the is_data_available test works in hover.py and in box.py. This predicate function returns true when all sensor data is available. the previous version might have been causing a bug with a physical drone.

Add SetpointSender. This object sends setpoints at a set frequency. 

Add getter function for  SensorData to SensorSynchronizer. This function lets us read the current position in the box test. We need the current position to calculate our way points.