Add a method to find the gimbal managers
To find the gimbal managers we need to broadcasts a mavlink message that requests all gimbal managers to respond with their details.

Add a method to get the gimbal managers we've previously found. Every time we receive information about a gimbal manager we record it in a data structure. This method just returns all the gimbal managers we've found.

WIP #4