Add test for the RC failsafe
This test walks the user through every RC input that must trigger the program to exit in an emergency. It also tests if we can detect the RC input that signals when the user wants to start a test.

We need a test to make sure that our hardware tests only start when the user triggers it with the RC transmitter. We also need to test that we can detect all the RC inputs that signal the user is trying to take control of the drone in an emergency.

See GitHub issue #8 this comment