Don't wait for RC tx data on init
Before flying, the program does pre-flight checks. It makes sure that it can read all the data it needs. Among that data is the RC transmitter state. But when simulating, an RC transmitter might not be connected. This causes the program to crash.

This change removes the RC transmitter check from the is_data_available function.