Update docker-compose for SD-HX10
There's a Jetson on the  SD-HX10 and it's connected to the flight controller using /dev/ttyUSB0

This updates the docker-compose.yml file so that mavros can communicate with the flight controller. When we start mavros, we need to tell it how to reach the flight controller. The fcu_url parameter specifies what device or port to use.