Setting preflight parameters
The values that were chosen for setting parameters were correct. However, it could be dangerous for flight operations at a location when the crew set the PX4 parameters manually due to topography, etc., only to find that they are reset in the Jetson code. Proposed replacing set_preflight_params with check_preflight_params. Message will be sent to Jetson recommending the parameters be changed by the pilot.