Rework the check_preflight_params
When testing the function, sometimes the get_param function didn't work. So I reworked method to show an error message when that happens.

Also during testing, I found it was handy to see the value was in case there was a problem with it. I also changed the error messages so that a link to the docs is given for the parameter in case something is wrong.

With all these new features it made sense to refactored the tests to be data-oriented.  It was too hard to read otherwise.