Add read_takeoff_alt method and move sleep to new module
The hover test and the box test both need to read the takeoff alt. And it's a little tricky to do this. So this commit adds a function to the drone object that can read the takeoff parameter.

Also move the sleep function to it's own module. Needed to do this because the new  read_takeoff_alt function needed it. And that caused a circular import. So to get around the issue, the sleep function is in a new module called sleepy.