Remain armed for 9.5 seconds
Before this change, the drone would wait on the ground in the armed state for 10 seconds before taking off. This was a problem because the default Time-out for auto disarm if not taking off is 10 seconds (see the parameter called COM_DISARM_PRFLT). The program would send the arm command, then it would take some time for it to realize that the drone is armed, then it would start waiting for the specified amount of time. When this wait time was 10 seconds, we were guaranteed disarm before takeoff. Setting this to 9.5 seconds is cutting it pretty close, but hopefully it's an acceptable value.