Add method to send velocity setpoints
After we arm, we want to put the drone in offboard mode before we start the RC failsafe. To do this we must stream setpoints to the flight controller. The safest set point to stream seems to be a velocity set point where all values are zero.

A position setpoint is less safe because the position could drift over time and we don't want the drone to try and move while it's on the ground. 

So this commit adds a method to send velocity setpoints.

See GitHub Issue #12