Wait for the pilot to arm drone #13
We're concerned about accidentally disarming mid-flight. This seems likely in case the pilot takes control from the companion computer. This changes the hover and box tests so that they don't arm the drone automatically. Instead they wait for the pilot to arm the drone. 

When the companion computer arms the drone, the arm/disarm toggle switch on the SD-HX10's RC unit never leaves the disarm position. This has been fine so far because PX4 won't disarm while in offboard mode. But in some human-piloting modes like stabilized and manual mode PX4 will disarm mid-flight. By default, the drone will go into stabilized mode if the pilot takes control from the companion computer. So we need to make sure that switch is set to the armed position before takeoff.

See GitHub Issue #13