Add mavlink-router WIP #14
This was added so we can better troubleshoot in the field. During the last field test, we needed better visibility into what the system was doing. Something went wrong during the hover test and the log messages didn't provide enough detail. To address this, I added mavlink router. With this program, we can run QGroundControl and the hardware tests at the same time. mavlink-router accesses the serial port directly, then it forwards mavlink packets to and from other programs over the network (programs like mavros and QGroundControl). QGroundControl will give the ground operators visibility into PX4 and that will help us detect and resolve problems in the field.
 
Add a dockerfile and a config file for mavlink router