Work around docker bug #14
When I tested the dockerfile for mavlink router I ran into a bug in docker. I couldn't run apt-get update and it seems to be a problem with ubuntu 22.04 images and docker. I found this:

https://stackoverflow.com/questions/71941032/why-i-cannot-run-apt-update-inside-a-fresh-ubuntu22-04

To work around the problem I switch to ubuntu 20.04.  Because of this change, meson gets installed through pip (because the 20.04 repos have a version that's too old).