Switch mavros to use mavlink router WIP #14
Update the docker compose files for the jetson and the aero so that mavros connects to PX4 via the mavlink-router. Also update docker-compose files so that mavlink-router will access PX4 using the right serial device. Comment out the uart endpoint in the mavlink-router config file because the aero and the jetson use a different serial device and it makes sense for that detail to be specified in the docker compose file.