Add mavlink-router #14
This was added so we can better troubleshoot in the field. During the last field test, we needed better visibility into what the system was doing. Something went wrong during the hover test and the log messages didn't provide enough detail. To address this, I added mavlink router. With this program, we can run QGroundControl and the hardware tests at the same time. mavlink-router accesses the serial port directly, then it forwards mavlink packets to and from other programs over the network (programs like mavros and QGroundControl). QGroundControl will give the ground operators visibility into PX4 and that will help us detect and resolve problems in the field.
 
Add the mavlink_router service to the docker compose file. Add dockerfile and a config file for this service. 

Change mavros to use mavlink_router

Update the docker compose files for the jetson and the aero so that mavros connects to PX4 via mavlink_router. Also update docker-compose files so that mavlink-router will access PX4 using the correct serial device.

Specify that mavros depends on mavlink_router in the docker compose file

Fix baud rate in docker config file for Intel Aero