Add instructions for updating files in place
When running tests, the drone doesn't have an internet connection. For small changes, we need the ability to update files without running "docker-compose build." We can't use "docker-compose build" because that command requires an internet connection. So I added instructions for updating the files in a running container. 

Also add note about how the hover test and box test expect you to arm the drone with the RC controller.