Add dev_test service to docker-compose.yaml
This change build on the last change.

When running tests, the drone doesn't have an internet connection. For small changes, we need the ability to update files without running "docker-compose build." We can't use "docker-compose build" because that command requires an internet connection.

This adds a new service lets you make changes and run the tests in docker without having to re-build. The project files are mounted in the container.

This is an easier way to make small changes in the field.

Also fix some spelling typos.