Fix rapid yaw change during box test
The box test used to set the yaw value so that the drone faced East. This caused the drone to rapidly spin in the air. This fix changes the box test so that the starting yaw is preserved.

Clean up the code the reads the drone's position
Add a function to read the drone's compass heading
Add a methods to the SetpointSender to get or set the drone's yaw
Change the send_setpoint methods to accept a yaw argument