Change the takeoff sequence for the hover test
When testing, the motors spin too quickly while on the ground and in offboard mode.

To address this, the hover test now waits 8 seconds after arming then it enters offboard mode.
Once in offboard mode, the RC failsafe turns on and the aircraft waits on the ground for 2 more seconds.
Finally, the aircraft enters takeoff mode.

\#27